+++ R/mod_map.R
@@ -98,36 +98,5 @@ mod_map_server <- function(input, output, session, df){
           symbolSize = 20
         )
       )
-      
-  #   dat %>% 
-  #     dplyr::group_by(date) %>% 
-  #     echarts4r::e_charts(chinese, timeline = TRUE) %>% 
-  #     echarts4r.maps::em_map("China") %>% 
-  #     echarts4r::e_map(cases, map = "China", name = "confirmed") %>% 
-  #     echarts4r::e_visual_map(
-  #       cases, 
-  #       show = FALSE, 
-  #       min = 0,
-  #       orient = "horizontal",
-  #       right = "center"
-  #     ) %>% 
-  #     echarts4r::e_theme(theme) %>% 
-  #     echarts4r::e_tooltip() %>% 
-  #     echarts4r::e_timeline_opts(currentIndex = index) %>% 
-  #     echarts4r::e_timeline_serie(
-  #       title = titles
-  #     ) %>% 
-  #     echarts4r::e_timeline_opts(
-  #       playInterval = 600, 
-  #       symbolSize = 4, 
-  #       axis_type = "time",
-  #       label = list(
-  #         show = FALSE
-  #       ),
-  #       checkpointStyle = list(
-  #         symbol = "pin",
-  #         symbolSize = 20
-  #       )
-  #     )
   })
 }