+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Add hubei default province in DXY tab.
 - Add shinyscroll to scroll to selected province in DXY tab.
+- Extend toast duration to 5 seconds.
 
 # coronavirus 0.0.4